requirements.txt
Django==1.9.4
django-bootstrap-form==3.2
dj-database-url==0.4.0
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
python-social-auth==0.2.14
django-suit==0.2.18
djangorestframework==3.3.3
Pillow==2.7
names